remove folder and file facades
